Fix UTF-8 string size in 'read_node'
'sord_node_get_string_counted' returns character count, which then is used
to forge strings and literals. But the character count is different from
raw string size for strings containing multi-byte UTF-8 characters. Better
use 'strlen' to get the raw byte size of the strings to be forged,
if not, strings with multi-byte UTF-8 characters are truncatedly forged.

To make this commit as fun as possible, klingon characters have been
used to extend the unit test.